Refactor home_page.dart to reduce file size from 3211 to 2569 lines
This commit refactors the overly large home_page.dart file by extracting
reusable components into separate files, improving code organization and
maintainability.

Changes:
- Extracted dialog widgets:
  - SortDialog: Handles sorting options
  - DateFilterDialog: Manages date range filtering
  - ReminderFilterDialog: Filters notes by reminder status
  - CategoryFilterDialog: Filters notes by category

- Created utility classes:
  - DateFormatter: Centralizes date formatting logic
  - SortType: Enum with sorting logic for notes
  - AttachmentCacheService: Manages attachment count caching

- Simplified HomePage:
  - Reduced from 3211 to 2569 lines (20% reduction)
  - Removed duplicate date formatting methods
  - Replaced inline dialog code with reusable widgets
  - Improved code readability and maintainability

Benefits:
- Better code organization with single responsibility principle
- Easier to test individual components
- Reduced complexity in HomePage
- Reusable widgets for future use
